Keep validator-statistics.groovy as is and add extended-validator-statistics.groovy
The latter deals with problematic paths and known irregularities.
